refactor: 주소 스크린와 주소 검색 스크린 UI 및 기능 분리
