+++ src/navigators/MainNavigator.tsx
@@ -3,6 +3,7 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import useAuth from 'src/hooks/useAuth';
 import Address from 'src/screens/Address/Address';
+import AddressSearch from 'src/screens/AddressSearch/AddressSearch';
 import Detail from 'src/screens/Detail/Detail';
 import Login from 'src/screens/Login/Login';
 import Main from 'src/screens/Main/Main';
@@ -56,6 +57,13 @@ const MainNavigator = () => {
           <Stack.Screen
             name="Address"
             component={Address}
+            options={{
+              headerShown: false,
+            }}
+          />
+          <Stack.Screen
+            name="AddressSearch"
+            component={AddressSearch}
             options={{
               headerShown: false,
             }}
