+++ src/navigators/types.ts
@@ -7,6 +7,8 @@ export type RootStackParamList = {
   };
   Search: undefined;
   Address: undefined;
+  AddressSearch: undefined;
+  AddressLocation: undefined;
   Detail: { cafeId: string };
   Map: undefined;
   Profile: undefined;
